+++ packages/blade/src/components/SpotlightPopoverTour/TourPopover.web.tsx
@@ -117,6 +117,7 @@ const TourPopover = ({
       close,
       defaultInitialFocusRef,
       titleId,
+      openInteraction: 'click' as const,
     };
   }, [close, titleId]);